refactor(controller): move engine-specific helpers out of backend-builder into their specs
backend-builder advertised itself as engine-agnostic while carrying the
vLLM extra-args allowlist filter, the vLLM Docker JIT wrapper, and the
llama.cpp binary resolution/serialization rules. Those now live in
specs/vllm-spec.ts and specs/llamacpp-spec.ts where the EngineSpec
docstring says they belong; backend-builder keeps the generic
serialization seam (appendSerializedArguments + ExtraArgumentSerializer)
and the shared docker-run shape. Adding an engine no longer means
editing the shared builder.
