+++ controller/src/modules/engines/specs/llamacpp-spec.ts
@@ -10,9 +10,9 @@ import type {
 } from "@local-studio/contracts/system";
 import { getLlamacppRuntimeInfo } from "../runtimes/runtime-info";
 import {
-  appendLlamacppArguments,
+  appendSerializedArguments,
   getExtraArgument,
-  resolveLlamaBinary,
+  type ExtraArgumentSerializer,
 } from "../process/backend-builder";
 import { stripForeignFlagKeys } from "@local-studio/contracts/engine-args";
 import { extractFlag } from "../argument-utilities";
@@ -22,7 +22,54 @@ import {
   LLAMACPP_UPGRADE_ENV,
   runEnvironmentUpgradeCommand,
 } from "../runtimes/upgrade-config";
-import { installManagedLlamacpp } from "../runtimes/managed-llamacpp";
+import { installManagedLlamacpp, managedLlamaServerPath } from "../runtimes/managed-llamacpp";
+
+const executableBaseName = (value: string): string => {
+  return value.split(/[\\/]/).filter(Boolean).at(-1)?.toLowerCase() ?? value.toLowerCase();
+};
+const isAllowedLlamaServerBinary = (value: string): boolean => {
+  const name = executableBaseName(value);
+  return name === "llama-server" || name === "llama-server.exe";
+};
+const rejectPathTraversal = (value: string, label: string): void => {
+  if (value.split(/[\\/]+/).includes("..")) {
+    throw new Error(`Invalid ${label}: path traversal is not allowed`);
+  }
+};
+
+export const resolveLlamaBinary = (recipe: Recipe, config: Config): string => {
+  const override = getExtraArgument(recipe.extra_args, "llama_bin") ?? config.llama_bin;
+  if (typeof override === "string" && override.trim()) {
+    rejectPathTraversal(override, "llama_bin");
+    if (!isAllowedLlamaServerBinary(override)) {
+      throw new Error("Invalid llama_bin: only llama-server executables are allowed");
+    }
+    const resolved = resolveBinary(override);
+    if (resolved) {
+      return resolved;
+    }
+    throw new Error(`Invalid llama_bin: executable "${override}" was not found`);
+  }
+  const managed = managedLlamaServerPath(config);
+  return resolveBinary("llama-server") ?? (existsSync(managed) ? managed : "llama-server");
+};
+
+const serializeLlamacppArgument: ExtraArgumentSerializer = (flag, _key, value) => {
+  if (value === true) return [flag];
+  if (value === false || value === undefined || value === null || value === "") return [];
+  if (Array.isArray(value)) {
+    return value.flatMap((entry) =>
+      entry === undefined || entry === null || entry === "" ? [] : [flag, String(entry)],
+    );
+  }
+  if (typeof value === "object") return [flag, JSON.stringify(value)];
+  return [flag, String(value)];
+};
+
+export const appendLlamacppArguments = (
+  command: string[],
+  extraArguments: Record<string, unknown>,
+): string[] => appendSerializedArguments(command, extraArguments, serializeLlamacppArgument);
 
 export const buildLlamacppRecipeArguments = (recipe: Recipe): string[] => {
   const command: string[] = [];
